backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
apscheduler
aiofiles
python-dateutil
pytz
jinja2
